split VolumeSplitN into its own util package
